Removed multiserver examples, and renamed nothreads to multiserver.
